Refactor: 차트에서 멀티레인저 분리
