+++ FE/src/components/PriceChart/index.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 
-import Chart from '@/components/Chart';
-import { calculateXRatio, calculateYRatio } from '@/components/Chart/util';
+import Chart from '@/components/PriceChart/Chart';
 import { CANVAS_HEIGHT, CANVAS_WIDTH } from '@/components/PriceChart/constants';
+import MultiRangerSlider from '@/components/PriceChart/MultiRangerSlider';
 
 import * as S from './style';
 
@@ -28,32 +28,6 @@ export default function PriceChart({ chartInfo, axis, xStep, yStep }: PriceChart
   const [leftThumbX, setLeftThumbX] = useState(0);
   const [rightThumbX, setRightThumbX] = useState(maximumX);
 
-  // FIXME: handleMaxPriceChange와 handleMinPriceChange 내부 로직이 비슷한 흐름인데 중복을 줄일 수 없을까?
-  const handleMaxPriceChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const {
-      target: { value },
-    } = event;
-    const newValue = Number(value);
-    if (newValue <= leftThumbX) {
-      setLeftThumbX(newValue - xStep);
-    }
-    setRightThumbX(newValue);
-  };
-
-  const handleMinPriceChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const {
-      target: { value },
-    } = event;
-    const newValue = Number(value);
-    if (newValue >= rightThumbX) {
-      setRightThumbX(newValue + xStep);
-    }
-    setLeftThumbX(newValue);
-  };
-
-  const moveLeftThumbX = (leftThumbX / maximumX) * 100;
-  const moveRightThumbX = 100 - (rightThumbX / maximumX) * 100;
-
   return (
     <S.Container containerWidth={CANVAS_HEIGHT} containerHeight={CANVAS_WIDTH}>
       <Chart
@@ -64,28 +38,14 @@ export default function PriceChart({ chartInfo, axis, xStep, yStep }: PriceChart
         leftThumbX={leftThumbX}
         rightThumbX={rightThumbX}
       />
-      <S.SliderController
-        type="range"
-        value={leftThumbX}
-        step={xStep}
-        min={0}
-        max={maximumX}
-        onChange={handleMinPriceChange}
+      <MultiRangerSlider
+        leftThumbX={leftThumbX}
+        rightThumbX={rightThumbX}
+        setLeftThumbX={setLeftThumbX}
+        setRightThumbX={setRightThumbX}
+        xStep={xStep}
+        maximumX={maximumX}
       />
-      <S.SliderController
-        type="range"
-        value={rightThumbX}
-        step={xStep}
-        min={0}
-        max={maximumX}
-        onChange={handleMaxPriceChange}
-      />
-      <S.VirtualSlider>
-        <S.Track></S.Track>
-        <S.Range moveLeftThumbX={moveLeftThumbX} moveRightThumbX={moveRightThumbX}></S.Range>
-        <S.LeftThumb moveLeftThumbX={moveLeftThumbX}></S.LeftThumb>
-        <S.RightThumb moveRightThumbX={moveRightThumbX}></S.RightThumb>
-      </S.VirtualSlider>
     </S.Container>
   );
 }
